refactor(backup): extract fs primitives into backup/fs_util.rs
